fix: resolve typos and seperate out test
