+++ src/models/transactions/checkCancel.ts
@@ -1,7 +1,7 @@
 import { ValidationError } from "../../common/errors";
-import { CommonFields, verifyCommonFields } from "./common";
+import { BaseTransaction, verifyBaseTransaction } from "./common";
 
-export interface CheckCancel extends CommonFields {
+export interface CheckCancel extends BaseTransaction {
     TransactionType: "CheckCancel";
     CheckID: string;
 }
@@ -14,7 +14,7 @@ export interface CheckCancel extends CommonFields {
  * @throws - When the CheckCancel is Malformed.
  */
  export function verifyCheckCancel(tx: CheckCancel): void {
-    verifyCommonFields(tx)
+    verifyBaseTransaction(tx)
 
     if (tx.CheckID !== undefined && typeof tx.CheckID !== 'string')
         throw new ValidationError("CheckCancel: invalid CheckID")
